pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nature_remo_con"]

[project]
name = "nature-remo-con"
version = "0.1.2"
description = "Nature Remo E で潮流を監視して TP-Link スマートプラグを制御する"
requires-python = ">=3.12.4"
dependencies = [
    "python-kasa",
    "requests",
    "toml",
    "netifaces",
    "python-dotenv>=1.1.0",
    "loguru>=0.7.3",
]

classifiers = ["Operating System :: Microsoft :: Windows :: Windows 11"]

[tool.pytest.ini_options]
addopts = "-ra -q -v"
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[project.optional-dependencies]
dev = ["pytest>=8.3.3", "pytest-cov>=5.0.0", "pytest-asyncio>=0.24.0"]

[project.scripts]
nature-remo-con = "nature_remo_con.main:cli"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0"
]
